+++ plugin-dev/Source/Sentry/Private/Android/AndroidSentryScope.h
@@ -18,8 +18,9 @@ public:
 	virtual void ClearBreadcrumbs() override;
 	virtual void AddAttachment(TSharedPtr<ISentryAttachment> attachment) override;
 	virtual void ClearAttachments() override;
-	virtual void SetTagValue(const FString& key, const FString& value) override;
-	virtual FString GetTagValue(const FString& key) const override;
+	virtual void SetTag(const FString& key, const FString& value) override;
+	virtual FString GetTag(const FString& key) const override;
+	virtual bool TryGetTag(const FString& key, FString& value) const override;
 	virtual void RemoveTag(const FString& key) override;
 	virtual void SetTags(const TMap<FString, FString>& tags) override;
 	virtual TMap<FString, FString> GetTags() const override;
@@ -27,13 +28,16 @@ public:
 	virtual TArray<FString> GetFingerprint() const override;
 	virtual void SetLevel(ESentryLevel level) override;
 	virtual ESentryLevel GetLevel() const override;
-	virtual void SetContext(const FString& key, const TMap<FString, FString>& values) override;
+	virtual void SetContext(const FString& key, const TMap<FString, FSentryVariant>& values) override;
+	virtual TMap<FString, FSentryVariant> GetContext(const FString& key) const override;
+	virtual bool TryGetContext(const FString& key, TMap<FString, FSentryVariant>& value) const override;
 	virtual void RemoveContext(const FString& key) override;
-	virtual void SetExtraValue(const FString& key, const FString& value) override;
-	virtual FString GetExtraValue(const FString& key) const override;
+	virtual void SetExtra(const FString& key, const FSentryVariant& value) override;
+	virtual FSentryVariant GetExtra(const FString& key) const override;
+	virtual bool TryGetExtra(const FString& key, FSentryVariant& value) const override;
 	virtual void RemoveExtra(const FString& key) override;
-	virtual void SetExtras(const TMap<FString, FString>& extras) override;
-	virtual TMap<FString, FString> GetExtras() const override;
+	virtual void SetExtras(const TMap<FString, FSentryVariant>& extras) override;
+	virtual TMap<FString, FSentryVariant> GetExtras() const override;
 	virtual void Clear() override;
 
 private: