+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentryScope.cpp
@@ -47,17 +47,26 @@ void FGenericPlatformSentryScope::ClearAttachments()
 	// Not available for generic platform
 }
 
-void FGenericPlatformSentryScope::SetTagValue(const FString& key, const FString& value)
+void FGenericPlatformSentryScope::SetTag(const FString& key, const FString& value)
 {
 	Tags.Add(key, value);
 }
 
-FString FGenericPlatformSentryScope::GetTagValue(const FString& key) const
+FString FGenericPlatformSentryScope::GetTag(const FString& key) const
 {
 	if (!Tags.Contains(key))
 		return FString();
 
 	return Tags[key];
+}
+
+bool FGenericPlatformSentryScope::TryGetTag(const FString& key, FString& value) const
+{
+	if (!Tags.Contains(key))
+		return false;
+
+	value = Tags[key];
+	return true;
 }
 
 void FGenericPlatformSentryScope::RemoveTag(const FString& key)
@@ -95,9 +104,26 @@ ESentryLevel FGenericPlatformSentryScope::GetLevel() const
 	return Level;
 }
 
-void FGenericPlatformSentryScope::SetContext(const FString& key, const TMap<FString, FString>& values)
+void FGenericPlatformSentryScope::SetContext(const FString& key, const TMap<FString, FSentryVariant>& values)
 {
 	Contexts.Add(key, values);
+}
+
+TMap<FString, FSentryVariant> FGenericPlatformSentryScope::GetContext(const FString& key) const
+{
+	if (!Contexts.Contains(key))
+		return TMap<FString, FSentryVariant>();
+
+	return Contexts[key];
+}
+
+bool FGenericPlatformSentryScope::TryGetContext(const FString& key, TMap<FString, FSentryVariant>& value) const
+{
+	if (!Contexts.Contains(key))
+		return false;
+
+	value = Contexts[key];
+	return true;
 }
 
 void FGenericPlatformSentryScope::RemoveContext(const FString& key)
@@ -108,17 +134,26 @@ void FGenericPlatformSentryScope::RemoveContext(const FString& key)
 	Contexts.Remove(key);
 }
 
-void FGenericPlatformSentryScope::SetExtraValue(const FString& key, const FString& value)
+void FGenericPlatformSentryScope::SetExtra(const FString& key, const FSentryVariant& value)
 {
 	Extra.Add(key, value);
 }
 
-FString FGenericPlatformSentryScope::GetExtraValue(const FString& key) const
+FSentryVariant FGenericPlatformSentryScope::GetExtra(const FString& key) const
 {
 	if (!Extra.Contains(key))
-		return FString();
+		return FSentryVariant();
 
 	return Extra[key];
+}
+
+bool FGenericPlatformSentryScope::TryGetExtra(const FString& key, FSentryVariant& value) const
+{
+	if (!Extra.Contains(key))
+		return false;
+
+	value = Extra[key];
+	return true;
 }
 
 void FGenericPlatformSentryScope::RemoveExtra(const FString& key)
@@ -129,12 +164,12 @@ void FGenericPlatformSentryScope::RemoveExtra(const FString& key)
 	Extra.Remove(key);
 }
 
-void FGenericPlatformSentryScope::SetExtras(const TMap<FString, FString>& extras)
+void FGenericPlatformSentryScope::SetExtras(const TMap<FString, FSentryVariant>& extras)
 {
 	Extra.Append(extras);
 }
 
-TMap<FString, FString> FGenericPlatformSentryScope::GetExtras() const
+TMap<FString, FSentryVariant> FGenericPlatformSentryScope::GetExtras() const
 {
 	return Extra;
 }
@@ -172,12 +207,12 @@ void FGenericPlatformSentryScope::Apply(sentry_scope_t* scope)
 
 	for (const auto& ExtraItem : Extra)
 	{
-		sentry_scope_set_extra(scope, TCHAR_TO_UTF8(*ExtraItem.Key), sentry_value_new_string(TCHAR_TO_UTF8(*ExtraItem.Value)));
+		sentry_scope_set_extra(scope, TCHAR_TO_UTF8(*ExtraItem.Key), FGenericPlatformSentryConverters::VariantToNative(ExtraItem.Value));
 	}
 
 	for (const auto& ContextsItem : Contexts)
 	{
-		sentry_scope_set_context(scope, TCHAR_TO_UTF8(*ContextsItem.Key), FGenericPlatformSentryConverters::StringMapToNative(ContextsItem.Value));
+		sentry_scope_set_context(scope, TCHAR_TO_UTF8(*ContextsItem.Key), FGenericPlatformSentryConverters::VariantMapToNative(ContextsItem.Value));
 	}
 
 	sentry_scope_set_level(scope, FGenericPlatformSentryConverters::SentryLevelToNative(Level));