+++ plugin-dev/Source/Sentry/Private/GenericPlatform/Infrastructure/GenericPlatformSentryConverters.cpp
@@ -197,59 +197,81 @@ FSentryVariant FGenericPlatformSentryConverters::VariantToUnreal(sentry_value_t 
 	case SENTRY_VALUE_TYPE_STRING:
 		return FSentryVariant(FString(sentry_value_as_string(variant)));
 	case SENTRY_VALUE_TYPE_LIST:
-	{
-		TArray<FSentryVariant> unrealArray;
-
-		int32 len = sentry_value_get_length(variant);
-		for (int32 i = 0; i < len; ++i)
-		{
-			unrealArray.Add(VariantToUnreal(sentry_value_get_by_index(variant, i)));
-		}
-
-		return FSentryVariant(unrealArray);
-	}
+		return VariantArrayToUnreal(variant);
 	case SENTRY_VALUE_TYPE_OBJECT:
-	{
-		TMap<FString, FSentryVariant> unrealMap;
-
-		FString mapJsonString = FString(sentry_value_to_json(variant));
-		if (mapJsonString.IsEmpty() || mapJsonString.Equals(TEXT("null")))
-		{
-			return unrealMap;
-		}
-
-		TSharedPtr<FJsonObject> jsonObject;
-		TSharedRef<TJsonReader<>> jsonReader = TJsonReaderFactory<>::Create(mapJsonString);
-		bool bDeserializeSuccess = FJsonSerializer::Deserialize(jsonReader, jsonObject);
-		if (!bDeserializeSuccess)
-		{
-			UE_LOG(LogSentrySdk, Error, TEXT("VariantToUnreal failed to deserialize map Json."));
-			return unrealMap;
-		}
-
-		TArray<FString> keysArr;
-		jsonObject->Values.GetKeys(keysArr);
-
-		for (auto it = keysArr.CreateConstIterator(); it; ++it)
-		{
-			unrealMap.Add(*it, VariantToUnreal(sentry_value_get_by_key(variant, TCHAR_TO_ANSI(**it))));
-		}
-
-		return FSentryVariant(unrealMap);
-	}
+		return VariantMapToUnreal(variant);
 	default:
 		UE_LOG(LogSentrySdk, Warning, TEXT("Unknown Sentry variant type used - an empty one will be returned."));
 		return FSentryVariant();
 	}
 }
 
+TMap<FString, FSentryVariant> FGenericPlatformSentryConverters::VariantMapToUnreal(sentry_value_t map)
+{
+	TMap<FString, FSentryVariant> unrealMap;
+
+	char* jsonString = sentry_value_to_json(map);
+	if (!jsonString)
+	{
+		return unrealMap;
+	}
+
+	FString mapJsonString = FString(jsonString);
+	if (mapJsonString.IsEmpty() || mapJsonString.Equals(TEXT("null")))
+	{
+		sentry_string_free(jsonString);
+		return unrealMap;
+	}
+
+	TSharedPtr<FJsonObject> jsonObject;
+	TSharedRef<TJsonReader<>> jsonReader = TJsonReaderFactory<>::Create(mapJsonString);
+	bool bDeserializeSuccess = FJsonSerializer::Deserialize(jsonReader, jsonObject);
+	if (!bDeserializeSuccess)
+	{
+		UE_LOG(LogSentrySdk, Error, TEXT("VariantToUnreal failed to deserialize map Json."));
+		sentry_string_free(jsonString);
+		return unrealMap;
+	}
+
+	TArray<FString> keysArr;
+	jsonObject->Values.GetKeys(keysArr);
+
+	for (auto it = keysArr.CreateConstIterator(); it; ++it)
+	{
+		unrealMap.Add(*it, VariantToUnreal(sentry_value_get_by_key(map, TCHAR_TO_ANSI(**it))));
+	}
+
+	sentry_string_free(jsonString);
+	return unrealMap;
+}
+
+TArray<FSentryVariant> FGenericPlatformSentryConverters::VariantArrayToUnreal(sentry_value_t array)
+{
+	TArray<FSentryVariant> unrealArray;
+
+	int32 len = sentry_value_get_length(array);
+	for (int32 i = 0; i < len; ++i)
+	{
+		unrealArray.Add(VariantToUnreal(sentry_value_get_by_index(array, i)));
+	}
+
+	return unrealArray;
+}
+
 TMap<FString, FString> FGenericPlatformSentryConverters::StringMapToUnreal(sentry_value_t map)
 {
 	TMap<FString, FString> unrealMap;
 
-	FString mapJsonString = FString(sentry_value_to_json(map));
+	char* jsonString = sentry_value_to_json(map);
+	if (!jsonString)
+	{
+		return unrealMap;
+	}
+
+	FString mapJsonString = FString(jsonString);
 	if (mapJsonString.IsEmpty() || mapJsonString.Equals(TEXT("null")))
 	{
+		sentry_string_free(jsonString);
 		return unrealMap;
 	}
 
@@ -259,6 +281,7 @@ TMap<FString, FString> FGenericPlatformSentryConverters::StringMapToUnreal(sentr
 	if (!bDeserializeSuccess)
 	{
 		UE_LOG(LogSentrySdk, Error, TEXT("StringMapToUnreal failed to deserialize map Json."));
+		sentry_string_free(jsonString);
 		return unrealMap;
 	}
 
@@ -270,6 +293,7 @@ TMap<FString, FString> FGenericPlatformSentryConverters::StringMapToUnreal(sentr
 		unrealMap.Add(*it, jsonObject->GetStringField(*it));
 	}
 
+	sentry_string_free(jsonString);
 	return unrealMap;
 }
 
@@ -277,24 +301,10 @@ TArray<FString> FGenericPlatformSentryConverters::StringArrayToUnreal(sentry_val
 {
 	TArray<FString> unrealArray;
 
-	FString arrayJsonString = FString(sentry_value_to_json(array));
-	if (arrayJsonString.IsEmpty() || arrayJsonString.Equals(TEXT("null")))
+	int32 len = sentry_value_get_length(array);
+	for (int32 i = 0; i < len; ++i)
 	{
-		return unrealArray;
-	}
-
-	TArray<TSharedPtr<FJsonValue>> jsonArray;
-	TSharedRef<TJsonReader<>> jsonReader = TJsonReaderFactory<>::Create(arrayJsonString);
-	bool bDeserializeSuccess = FJsonSerializer::Deserialize(jsonReader, jsonArray);
-	if (!bDeserializeSuccess)
-	{
-		UE_LOG(LogSentrySdk, Error, TEXT("StringArrayToUnreal failed to deserialize array Json."));
-		return unrealArray;
-	}
-
-	for (auto it = jsonArray.CreateConstIterator(); it; ++it)
-	{
-		unrealArray.Add(it->Get()->AsString());
+		unrealArray.Add(sentry_value_as_string(sentry_value_get_by_index(array, i)));
 	}
 
 	return unrealArray;
@@ -353,6 +363,68 @@ ELogVerbosity::Type FGenericPlatformSentryConverters::SentryLevelToLogVerbosity(
 	}
 
 	return LogVerbosity;
+}
+
+TSharedPtr<FJsonValue> FGenericPlatformSentryConverters::VariantToJsonValue(const FSentryVariant& variant)
+{
+	switch (variant.GetType())
+	{
+	case ESentryVariantType::Integer:
+		return MakeShareable(new FJsonValueNumber(variant.GetValue<int32>()));
+	case ESentryVariantType::Float:
+		return MakeShareable(new FJsonValueNumber(variant.GetValue<float>()));
+	case ESentryVariantType::Bool:
+		return MakeShareable(new FJsonValueBoolean(variant.GetValue<bool>()));
+	case ESentryVariantType::String:
+		return MakeShareable(new FJsonValueString(variant.GetValue<FString>()));
+	case ESentryVariantType::Array:
+		return VariantArrayToJsonValue(variant.GetValue<TArray<FSentryVariant>>());
+	case ESentryVariantType::Map:
+		return VariantMapToJsonValue(variant.GetValue<TMap<FString, FSentryVariant>>());
+	default:
+		return MakeShareable(new FJsonValueNull());
+	}
+}
+
+TSharedPtr<FJsonValue> FGenericPlatformSentryConverters::VariantArrayToJsonValue(const TArray<FSentryVariant>& array)
+{
+	TArray<TSharedPtr<FJsonValue>> jsonArray;
+
+	for (auto it = array.CreateConstIterator(); it; ++it)
+	{
+		const FSentryVariant& variant = *it;
+		TSharedPtr<FJsonValue> jsonValue = VariantToJsonValue(variant);
+		if (jsonValue.IsValid())
+		{
+			jsonArray.Add(jsonValue);
+		}
+	}
+
+	return MakeShareable(new FJsonValueArray(jsonArray));
+}
+
+TSharedPtr<FJsonValue> FGenericPlatformSentryConverters::VariantMapToJsonValue(const TMap<FString, FSentryVariant>& map)
+{
+	TSharedPtr<FJsonObject> jsonObject = MakeShareable(new FJsonObject);
+
+	for (auto it = map.CreateConstIterator(); it; ++it)
+	{
+		const FString& key = it.Key();
+		const FSentryVariant& variant = it.Value();
+
+		if (variant.GetType() == ESentryVariantType::Empty)
+		{
+			continue;
+		}
+
+		TSharedPtr<FJsonValue> jsonValue = VariantToJsonValue(variant);
+		if (jsonValue.IsValid())
+		{
+			jsonObject->SetField(key, jsonValue);
+		}
+	}
+
+	return MakeShareable(new FJsonValueObject(jsonObject));
 }
 
 #endif