+++ plugin-dev/Source/Sentry/Private/Null/NullSentrySubsystem.h
@@ -14,7 +14,7 @@ public:
 	virtual bool IsEnabled() override { return false; }
 	virtual ESentryCrashedLastRun IsCrashedLastRun() override { return ESentryCrashedLastRun::NotEvaluated; }
 	virtual void AddBreadcrumb(TSharedPtr<ISentryBreadcrumb> breadcrumb) override {}
-	virtual void AddBreadcrumbWithParams(const FString& Message, const FString& Category, const FString& Type, const TMap<FString, FString>& Data, ESentryLevel Level) override {}
+	virtual void AddBreadcrumbWithParams(const FString& Message, const FString& Category, const FString& Type, const TMap<FString, FSentryVariant>& Data, ESentryLevel Level) override {}
 	virtual void ClearBreadcrumbs() override {}
 	virtual TSharedPtr<ISentryId> CaptureMessage(const FString& message, ESentryLevel level) override { return nullptr; }
 	virtual TSharedPtr<ISentryId> CaptureMessageWithScope(const FString& message, ESentryLevel level, const FSentryScopeDelegate& onConfigureScope) override { return nullptr; }
@@ -25,7 +25,7 @@ public:
 	virtual void SetUser(TSharedPtr<ISentryUser> user) override {}
 	virtual void RemoveUser() override {}
 	virtual void ConfigureScope(const FSentryScopeDelegate& onConfigureScope) override {}
-	virtual void SetContext(const FString& key, const TMap<FString, FString>& values) override {}
+	virtual void SetContext(const FString& key, const TMap<FString, FSentryVariant>& values) override {}
 	virtual void SetTag(const FString& key, const FString& value) override {}
 	virtual void RemoveTag(const FString& key) override {}
 	virtual void SetLevel(ESentryLevel level) override {}