+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+### Breaking Changes
+
+- "Value" suffix has been removed from the get/set function names for tags and extras in `SentryScope`.
+- Type of input parameters in certain public API functions was changed from `FString` to `FSentryVariant`:
+  - Get/Set functions for contexts and extras in `SentryScope`
+  - Get/Set functions for arbitrary data in `SentryBreadcrumb`
+  - Get/Set functions for arbitrary data in `SentryTransaction` and `SentrySpan`
+  - `AddBreadcrumbWithParams` and `SentContext` functions in `SentrySubsystem`
+  - `CreateSentryBreadcrumb` function in `SentryLibrary`
+
+### Features
+
+- Adopt generic variant type in public APIs ([#971](https://github.com/getsentry/sentry-unreal/pull/971))
+
 ## 1.0.0-beta.3
 
 ### Breaking Changes